// src/app/generated/player-api/model/webhookSubscription.ts
/**
 * Player API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventType } from './eventType';


export interface WebhookSubscription { 
    id?: string;
    name?: string | null;
    callbackUri?: string | null;
    clientId?: string | null;
    clientSecret?: string | null;
    readonly clientSecretSet?: boolean;
    lastError?: string | null;
    eventTypes?: Array<EventType> | null;
}